Split karva_diagnostic::result into a directory module
Promote result.rs (537 lines) into a result/ module with three focused submodules: kind.rs (IndividualTestResultKind, TestResultKind), stats.rs (TestResultStats + Display + serde + tests), flaky.rs (FlakyTest, FlakyTestRecord, Display wrappers). mod.rs keeps TestRunResult and the public re-exports. Pure file move; behavior unchanged.
